prevent memory corruption
An USB device that provides more endpoints then the defined maximum
supported ones of the code base this will lead to a memory corruption.
The code does not check if it runs over the maximum defined endpoints
and write to the memory behind of it.
For the number of interfaces the check if the maximum is already reached
is present, for the endpoints it is missing.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>